Add NativeEventEmitter listeners only once
Initially, a listener was created on each render of the ActionToolbar component and was never removed, resulting in an increasing number of message prompts when adding events to calendar. Creating and removing the listener on every EventDetails-visit seems to negatively affect performance. Solution is to create the calendar events listener only once, when the BarteguidenApp component mounts for iOS, and removing it when it unmounts.

Resolves #84