Replace Navigator with NavigationExperimental
Attempt to make navigation more understandable. There are two types of navigation used: tabs and card stack. The tabs are simply the four views defined in /views (excluding Settings on android). On iOS, a route is pushed to the card stack only when browsing event details. On android, additionally the settings view is pushed to the card stack. In short, there are three possible routes in the card stack: Either one is browsing the content in a tab, an event, or the android settings view. These rotues are defined in routes.js.

The two navigation types are used differently on the two platforms. On iOS the "outermost" navigation is the tabs, while it is the card stack on android. This way, the tabs disappear on android when browsing an event, avoiding some possible confusion.

The currently viewed tab and route are reflected in the redux state.

Resolves #49